+++ generators/extend-config/index.js
@@ -1,32 +1,12 @@
-/*
-* License-Identifier: Apache-2.0
-*/
-const Generator = require('yeoman-generator');
-const configTransformers = require('./configTransformers');
+import * as Generator from "yeoman-generator";
+import * as configTransformers from "./configTransformers";
+import parseFabricaConfig from "../utils/parseFabricaConfig";
+import { FabricaConfigJson } from "../types/FabricaConfigJson";
+import { FabricaConfigExtended } from "../types/FabricaConfigExtended";
 
-const ValidateGeneratorPath = require.resolve('../validate');
+const ValidateGeneratorPath = require.resolve("../validate");
 
-class ExtendConfigGenerator extends Generator {
-  constructor(args, opts) {
-    super(args, opts);
-    this.argument('fabricaConfig', {
-      type: String,
-      required: true,
-      description: 'fabrica config file path',
-    });
-
-    this.composeWith(ValidateGeneratorPath, { arguments: [this.options.fabricaConfig] });
-  }
-
-  async writing() {
-    const fabricaConfigPath = `${this.env.cwd}/${this.options.fabricaConfig}`;
-    const json = this.fs.readJSON(fabricaConfigPath);
-    const transformedConfig = ExtendConfigGenerator.extendJsonConfig(json);
-    this.log(JSON.stringify(transformedConfig, undefined, 2));
-  }
-}
-
-ExtendConfigGenerator.extendJsonConfig = (json) => {
+const extendConfig = (json: FabricaConfigJson): FabricaConfigExtended => {
   const {
     networkSettings: networkSettingsJson,
     rootOrg: rootOrgJson,
@@ -39,8 +19,12 @@ ExtendConfigGenerator.extendJsonConfig = (json) => {
   const rootOrg = configTransformers.transformRootOrgConfig(rootOrgJson);
   const orgs = configTransformers.transformOrgConfigs(orgsJson);
   const channels = configTransformers.transformChannelConfigs(channelsJson, orgs);
-  const chaincodes = configTransformers.transformChaincodesConfig(chaincodesJson, channels);
   const networkSettings = configTransformers.transformNetworkSettings(networkSettingsJson);
+  const chaincodes = configTransformers.transformChaincodesConfig(
+    networkSettingsJson.fabricVersion,
+    chaincodesJson,
+    channels,
+  );
 
   return {
     networkSettings,
@@ -52,4 +36,25 @@ ExtendConfigGenerator.extendJsonConfig = (json) => {
   };
 };
 
-module.exports = ExtendConfigGenerator;
+class ExtendConfigGenerator extends Generator {
+  constructor(args: string[], opts: Generator.GeneratorOptions) {
+    super(args, opts);
+    this.argument("fabricaConfig", {
+      type: String,
+      required: true,
+      description: "fabrica config file path",
+    });
+
+    this.composeWith(ValidateGeneratorPath, { arguments: [this.options.fabricaConfig] });
+  }
+
+  async writing(): Promise<void> {
+    const fabricaConfigPath = `${this.env.cwd}/${this.options.fabricaConfig}`;
+    const json = parseFabricaConfig(this.fs.read(fabricaConfigPath));
+    const transformedConfig = extendConfig(json);
+    this.log(JSON.stringify(transformedConfig, undefined, 2));
+  }
+}
+
+export { extendConfig };
+export default ExtendConfigGenerator;