+++ generators/validate/index.js
@@ -1,64 +1,85 @@
-/*
-* License-Identifier: Apache-2.0
-*/
-const Generator = require('yeoman-generator');
-const SchemaValidator = require('jsonschema').Validator;
-const chalk = require('chalk');
-const Listener = require('../utils/listener');
-const utils = require('../utils/utils');
+import * as Generator from "yeoman-generator";
+import { Schema, Validator as SchemaValidator } from "jsonschema";
+import * as chalk from "chalk";
+import schema from "../../docs/schema.json";
+import * as config from "../config";
+import parseFabricaConfig from "../utils/parseFabricaConfig";
+import { FabricaConfigJson, NetworkSettingsJson, OrdererJson, OrgJson } from "../types/FabricaConfigJson";
+import * as _ from "lodash";
 
-const schema = require('../../docs/schema.json');
-const config = require('../config');
-
-const CheckUpdatesGeneratorType = require.resolve('../check-updates');
+const ListCompatibleUpdatesGeneratorType = require.resolve("../list-compatible-updates");
 
 const validationErrorType = {
-  CRITICAL: 'validation-critical',
-  ERROR: 'validation-error',
-  WARN: 'validation-warning',
+  CRITICAL: "validation-critical",
+  ERROR: "validation-error",
+  WARN: "validation-warning",
 };
 
 const validationCategories = {
-  CRITICAL: 'Critical',
-  GENERAL: 'General',
-  ORDERER: 'Orderer',
-  PEER: 'Peer',
-  VALIDATION: 'Schema validation',
+  CRITICAL: "Critical",
+  GENERAL: "General",
+  ORDERER: "Orderer",
+  PEER: "Peer",
+  VALIDATION: "Schema validation",
 };
 
-module.exports = class extends Generator {
-  constructor(args, opts) {
+interface Message {
+  category: string;
+  message: string;
+}
+
+class Listener {
+  readonly messages: Message[] = [];
+
+  onEvent(event: Message) {
+    this.messages.push(event);
+  }
+
+  getAllMessages() {
+    return this.messages;
+  }
+
+  count() {
+    return this.messages.length;
+  }
+}
+
+const getFullPathOf = (configFile: string, currentPath: string): string => `${currentPath}/${configFile}`;
+
+class ValidateGenerator extends Generator {
+  private readonly errors = new Listener();
+  private readonly warnings = new Listener();
+
+  constructor(args: string[], opts: Generator.GeneratorOptions) {
     super(args, opts);
 
-    this.argument('fabricaConfig', {
+    this.argument("fabricaConfig", {
       type: String,
       required: true,
-      description: 'fabrica config file path',
+      description: "fabrica config file path",
     });
 
     this.addListener(validationErrorType.CRITICAL, (event) => {
-      this.log(chalk.bold.bgRed('Critical error occured:'));
+      this.log(chalk.bold.bgRed("Critical error occured:"));
       this.log(chalk.bold(`- ${event.message}`));
-      this._printIfNotEmpty(this.listeners.error.getAllMessages(), chalk.red.bold('Errors found:'));
+      this._printIfNotEmpty(this.errors.getAllMessages(), chalk.red.bold("Errors found:"));
       process.exit(1);
     });
 
-    this.listeners = { error: new Listener(), warn: new Listener() };
+    this.addListener(validationErrorType.ERROR, (e) => this.errors.onEvent(e));
+    this.addListener(validationErrorType.WARN, (e) => this.warnings.onEvent(e));
 
-    this.addListener(validationErrorType.ERROR, (e) => this.listeners.error.onEvent(e));
-    this.addListener(validationErrorType.WARN, (e) => this.listeners.warn.onEvent(e));
-
-    this.composeWith(CheckUpdatesGeneratorType, { arguments: [this.options.compatible] });
+    this.composeWith(ListCompatibleUpdatesGeneratorType);
   }
 
-  initializing() {
+  async initializing() {
     this.log(config.splashScreen());
   }
 
   async validate() {
     this._validateIfConfigFileExists(this.options.fabricaConfig);
 
-    const networkConfig = this.fs.readJSON(this.options.fabricaConfigPath);
+    const networkConfig = parseFabricaConfig(this.fs.read(this.options.fabricaConfigPath));
     this._validateJsonSchema(networkConfig);
     this._validateSupportedFabricaVersion(networkConfig.$schema);
     this._validateFabricVersion(networkConfig.networkSettings.fabricVersion);
@@ -69,26 +90,32 @@ module.exports = class extends Generator {
     this._validateOrgsAnchorPeerInstancesCount(networkConfig.orgs);
   }
 
-  async summary() {
-    this.log(`Errors count: ${this.listeners.error.count()}`);
-    this.log(`Warnings count: ${this.listeners.warn.count()}`);
-    this.log(chalk.bold('=================== Validation summary ==================='));
+  async shortSummary() {
+    this.log(`Validation errors count: ${this.errors.count()}`);
+    this.log(`Validation warnings count: ${this.warnings.count()}`);
+    this.log(chalk.bold("==========================================================="));
+  }
 
-    this._printIfNotEmpty(this.listeners.error.getAllMessages(), chalk.red.bold('Errors found:'));
-    this._printIfNotEmpty(this.listeners.warn.getAllMessages(), chalk.yellow('Warnings found:'));
+  async detailedSummary() {
+    const allValidationMessagesCount = this.errors.count() + this.warnings.count();
 
-    this.log(chalk.bold('==========================================================='));
+    if (allValidationMessagesCount > 0) {
+      this.log(chalk.bold("=================== Validation summary ===================="));
+      this._printIfNotEmpty(this.errors.getAllMessages(), chalk.red.bold("Errors found:"));
+      this._printIfNotEmpty(this.warnings.getAllMessages(), chalk.yellow("Warnings found:"));
+      this.log(chalk.bold("==========================================================="));
+    }
 
-    if (this.listeners.error.count() > 0) {
+    if (this.errors.count() > 0) {
       process.exit(1);
     }
   }
 
-  _validateIfConfigFileExists(configFilePath) {
-    const configFilePathAbsolute = utils.getFullPathOf(configFilePath, this.env.cwd);
+  _validateIfConfigFileExists(configFilePath: string) {
+    const configFilePathAbsolute = getFullPathOf(configFilePath, this.env.cwd);
     const fileExists = this.fs.exists(configFilePathAbsolute);
     if (!fileExists) {
-      const objectToEmit = {
+      const objectToEmit: Message = {
         category: validationCategories.CRITICAL,
         message: `No file under path: ${configFilePathAbsolute}`,
       };
@@ -98,9 +125,9 @@ module.exports = class extends Generator {
     }
   }
 
-  _validateJsonSchema(configToValidate) {
+  _validateJsonSchema(configToValidate: FabricaConfigJson) {
     const validator = new SchemaValidator();
-    const results = validator.validate(configToValidate, schema);
+    const results = validator.validate(configToValidate, schema as Schema);
     results.errors.forEach((result) => {
       const msg = `${result.property} : ${result.message}`;
       const objectToEmit = {
@@ -112,26 +139,26 @@ module.exports = class extends Generator {
     if (results.errors.length > 0) {
       const objectToEmit = {
         category: validationCategories.CRITICAL,
-        message: 'Json schema validation failed!',
+        message: "Json schema validation failed!",
       };
       this.emit(validationErrorType.CRITICAL, objectToEmit);
     }
   }
 
-  _printIfNotEmpty(messages, caption) {
+  _printIfNotEmpty(messages: Message[], caption: string) {
     if (messages.length > 0) {
       this.log(caption);
 
-      const grouped = utils.groupBy(messages, (msg) => msg.category);
+      const grouped: Record<string, Message[]> = _.groupBy(messages, (msg) => msg.category);
 
-      Array.from(grouped.keys()).forEach((key) => {
+      Object.keys(grouped).forEach((key) => {
         this.log(chalk.bold(`  ${key}:`));
-        grouped.get(key).forEach((msg) => this.log(`   - ${msg.message}`));
+        grouped[key].forEach((msg) => this.log(`   - ${msg.message}`));
       });
     }
   }
 
-  _validateSupportedFabricaVersion(schemaUrl) {
+  _validateSupportedFabricaVersion(schemaUrl: string) {
     const version = config.getVersionFromSchemaUrl(schemaUrl);
     if (!config.isFabricaVersionSupported(version)) {
       const msg = `Config file points to '${version}' Fabrica version which is not supported. Supported versions are: ${config.supportedVersionPrefix}x`;
@@ -143,7 +170,7 @@ module.exports = class extends Generator {
     }
   }
 
-  _validateFabricVersion(fabricVersion) {
+  _validateFabricVersion(fabricVersion: string) {
     if (!config.supportedFabricVersions.includes(fabricVersion)) {
       const objectToEmit = {
         category: validationCategories.GENERAL,
@@ -153,8 +180,8 @@ module.exports = class extends Generator {
     }
   }
 
-  _validateOrdererCountForSoloType(orderer) {
-    if (orderer.type === 'solo' && orderer.instances > 1) {
+  _validateOrdererCountForSoloType(orderer: OrdererJson) {
+    if (orderer.type === "solo" && orderer.instances > 1) {
       const objectToEmit = {
         category: validationCategories.ORDERER,
         message: `Orderer consesus type is set to 'solo', but number of instances is ${orderer.instances}. Only 1 instance will be created.`,
@@ -163,8 +190,8 @@ module.exports = class extends Generator {
     }
   }
 
-  _validateOrdererForRaftType(orderer, networkSettings) {
-    if (orderer.type === 'raft' || orderer.type === 'etcdraft') {
+  _validateOrdererForRaftType(orderer: OrdererJson, networkSettings: NetworkSettingsJson) {
+    if (orderer.type === "raft") {
       if (orderer.instances === 1) {
         const objectToEmit = {
           category: validationCategories.ORDERER,
@@ -191,12 +218,12 @@ module.exports = class extends Generator {
     }
   }
 
-  _validateOrgsAnchorPeerInstancesCount(orgs) {
+  _validateOrgsAnchorPeerInstancesCount(orgs: OrgJson[]) {
     orgs.forEach((org) => {
       const numberOfPeers = org.peer.instances;
       const numberOfAnchorPeers = org.peer.anchorPeerInstances;
 
-      if (numberOfPeers < numberOfAnchorPeers) {
+      if (!!numberOfAnchorPeers && numberOfPeers < numberOfAnchorPeers) {
         const objectToEmit = {
           category: validationCategories.PEER,
           message: `Too many anchor peers for organization "${org.organization.name}". Peer count: ${numberOfPeers}, anchor peer count: ${numberOfAnchorPeers}`,
@@ -205,4 +232,6 @@ module.exports = class extends Generator {
       }
     });
   }
-};
+}
+
+module.exports = ValidateGenerator;