+++ CHANGELOG.md
@@ -1,7 +1,17 @@
-## main (0.0.1)
+## main (0.1.0-unstable)
 
 ### Features
+* Support for private data ([#104](https://github.com/softwaremill/fabrica/issues/104))
+* Added channel query scripts  ([#169](https://github.com/softwaremill/fabrica/issues/169))
+* Support for Hyperledger Fabric 2.x (still with 1_4_x channel capabilities)
+* Support for fabrica config in YAML format
 
+### Chore & Maintenance
+* Rewrite Yeoman generators to use TypeScript
+
+## 0.0.1
+
+### Features
 * Generate simple Hyperledger Fabric network
 * Support for multiple organizations
 * Support for solo consensus protocol
@@ -9,14 +19,13 @@
 * Validation of `fabrica-config.json` based on JSON schema and other rules
 * Allow to upgrade chaincode ([#45](https://github.com/softwaremill/fabrica/issues/36))
 * Network recreation in one step ([#105](https://github.com/softwaremill/fabrica/issues/105))
-* Init command to provide simple config ([#90](https://github.com/softwaremill/fabrica/issues/90))
+* Init command to provide simple config ([#90](https://github.com/softwaremill/fabrica/issues/90)) with chaincode ([#100](https://github.com/softwaremill/fabrica/issues/100))  
+* Proper exposing Orderer and Peers ports for service discovery ([#116](https://github.com/softwaremill/fabrica/issues/116))  
 
 ### Fixes
-
 * Missing notify anchor peers step ([#26](https://github.com/softwaremill/fabrica/issues/26))
 
 ### Chore & Maintenance
-
 * Lint bash and YAML files ([#48](https://github.com/softwaremill/fabrica/issues/48))
 * Format generated scripts and YAML files ([#75](https://github.com/softwaremill/fabrica/issues/75))
 * Test Java chaincode on generated network ([#25](https://github.com/softwaremill/fabrica/issues/25))